shadow_removal/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
omegaconf>=2.1.0
swanlab>=0.3.0
Pillow>=8.0.0
numpy>=1.19.0
timm>=0.6.0
